Move xray test helpers into separate package (#2190)
* Rename testing.go to testing_test.go so that
"testing" is not imported in release code

* Move testing.go into its own xraytest pkg

* Change tests to refer to separate xraytest pkg

* Added package comment for staticcheck

(cherry picked from commit eeb68ca3a4fa4d7d821f932fe209ce764545bcbe)
